+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Upgraded Material Components For Android to 1.10.0
 - Fixed `DisplayMetrics.scaledDensity` deprecated in API level 34
 - Fixed Java "nullability problems" (using `@NonNull`)
+- Fixed Java "nullability and data flow problems" (using `Objects.requireNonNull`)
 
 
 ## [v1.2.7] (60) 㨆莊 (2023-11-05)